+++ resources/js/pages/Files/Browser.vue
@@ -14,9 +14,29 @@
                         <sui-icon name="add" class="purple corner" />
                     </i>
                 </sui-button>
+
+                <div is="sui-button-group" style="float: right">
+                    <sui-button class="icon" @click.native="toggleNewDir">
+                        <i class="icons" tooltip="New folder">
+                            <sui-icon name="folder outline" />
+                            <sui-icon name="add" class="purple corner" />
+                        </i>
+                    </sui-button>
+                </div>
             </path-bar>
 
             <file-list :files="files" :path="currentPath" />
+
+            <sui-modal size="tiny" v-model="promptNewDir">
+                <sui-modal-header>Enter folder name</sui-modal-header>
+                <sui-modal-content>
+                    <sui-input class="fluid" v-model="dirname" @keyup.enter="mkdir"
+                        style="border: 1px solid rgba(34, 36, 38, 0.15)" />
+                </sui-modal-content>
+                <sui-modal-actions>
+                    <sui-button positive @click.native="mkdir" content="OK" />
+                </sui-modal-actions>
+            </sui-modal>
 
             <sui-modal size="tiny" v-model="promptNewFile">
                 <sui-modal-header>Enter filename</sui-modal-header>
@@ -48,7 +68,9 @@ export default {
     },
     data() {
         return {
+            dirname: '',
             filename: '',
+            promptNewDir: false,
             promptNewFile: false,
         };
     },
@@ -78,6 +100,16 @@ export default {
                 name: 'files.edit',
                 query: { f: this.createPath },
             });
+        },
+        mkdir() {
+            const path = `${this.currentPath}/${this.dirname}`;
+
+            this.$store.dispatch('files/createDir', path).then(() => {
+                this.promptNewDir = false;
+            });
+        },
+        toggleNewDir() {
+            this.promptNewDir = !this.promptNewDir;
         },
         toggleNewFile() {
             this.promptNewFile = !this.promptNewFile;