+++ tests/Feature/Api/Files/CreatePathTest.php
@@ -30,7 +30,8 @@ class CreatePathTest extends TestCase
     {
         $file = resource_path('test-skel/created.txt');
 
-        $response = $this->authed()->postJson($this->endpoint(['file' => $file]), [
+        $response = $this->authed()->postJson($this->endpoint(), [
+            'file' => $file,
             'contents' => '# New File',
         ]);
 
@@ -38,5 +39,33 @@ class CreatePathTest extends TestCase
         $this->assertFileExists($file);
 
         unlink($file);
+    }
+
+    /** @test */
+    public function authed_user_can_create_a_folder(): void
+    {
+        $this->withoutExceptionHandling();
+        $dir = resource_path('test-skel/newdir');
+
+        $response = $this->authed()->postJson($this->endpoint(), [
+            'dir' => $dir,
+        ]);
+
+        $response->assertStatus(Response::HTTP_CREATED);
+        $this->assertDirectoryExists($dir);
+
+        rmdir($dir);
+    }
+
+    /** @test */
+    public function cannot_create_folder_if_target_exists(): void
+    {
+        $dir = resource_path('test-skel/mixed');
+
+        $response = $this->authed()->postJson($this->endpoint(), [
+            'dir' => $dir,
+        ]);
+
+        $response->assertStatus(Response::HTTP_CONFLICT);
     }
 }