+++ CHANGELOG.md
@@ -10,8 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Deleting system users and groups now triggers a confirmation prompt
 * Home directory can now be purged when deleting a system user
-* Files and folders can now be deleted through the File Manager
-* Files can be created in the file browser/viewer
+* File Manager now supports creating and deleting files and folders
+* Files can be renamed in the file browser
 
 ### Fixed
 * File contents no longer have whitespace and newlines trimmed when saved